pyproject.toml
[project]
name = "todo-app-backend-25-02-20"
version = "0.1.0"
description = ""
authors = [
    {name = "jen807",email = "80chill@naver.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
